[ERB Update] Upgrade to electron-react-boilerplate v2.0.1  (#34)
feat: upgrade entire codebase to use electron-react-boilerplate to v2.0.1 which includes the renaming of resources dir to assets, app to src, and puts configs, internals, and dll in .erb. Also adds notarization support and upgrades electron to 11 and webpack to 5. It also updates the workflows and includes other performance wins.

The app in development takes slightly longer to start now, but hot reloading is faster and works more consistently. Devs must use yarn start instead of yarn dev now.