Improve Emoji Display
Improving the display of emojis by using the awesome unicode-emoji package, which is basically a massive dataset of emojis and their description. We use it in the frontend to detect if an emoji version is older than 12.1. This change be changed later to detect what is on the system. Resolves #30